tratamento de exceção delete

+++ src/main/java/com/aprendendoSpring/couseSpring/services/UserService.java
@@ -3,8 +3,11 @@ package com.aprendendoSpring.couseSpring.services;
 import java.util.List;
 import java.util.Optional;
 
+import com.aprendendoSpring.couseSpring.services.exceptions.DatabaseException;
 import com.aprendendoSpring.couseSpring.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.aprendendoSpring.couseSpring.entities.User;
@@ -15,14 +18,14 @@ public class UserService {
 
 	@Autowired
 	private UserRepository repository;
-	
+
 	public List<User> findAll(){
 		return repository.findAll();
 	}
-	
+
 	public User findAllId(Long id) {
 		Optional<User> obj = repository.findById(id);
-		
+
 		return obj.orElseThrow(()-> new ResourceNotFoundException(id));
 	}
 
@@ -30,9 +33,17 @@ public class UserService {
 		return repository.save(obj);
 	}
 
-	public void delete(Long id){
-		repository.deleteById(id);
+	public void delete(Long id) {
+		try {
+			if(!repository.existsById(id)) {
+				throw new ResourceNotFoundException(id);
+			}
+			repository.deleteById(id);
+		}catch (DataIntegrityViolationException e){
+			throw new DatabaseException(e.getMessage());
+		}
 	}
+
 
 	public User update(Long id, User obj) {
 		User entity = repository.getReferenceById(id);
